requirements.txt
apache-airflow==2.6.0
dbt-core==1.5.0
apache-airflow-providers-snowflake==3.3.0
dbt-snowflake==1.5.0
